CDAP-4226 various changes to cleanup build
Changing the build to only produce a single jar instead of
batch and realtime jars, since the only difference between those
jars was their manifests. Instead, the single jar should be set
to extend both cdap-etl-batch and cdap-etl-realtime.

Also fixing widget configs to specify schema in the way the
3.3.x UI expects, and updating the cdap version to be
3.3.0-SNAPSHOT.

Finally, adding a build step that creates the config file for
the plugin, which specifies widget and doc information as
artifact properties.