CDAP-4415 copying cdap-etl-common from cdap as hydrator-common
there are three differences from cdap-etl-common:

The first is that RecordConverter was changed to use the cdap
Schema class rather than the Avro Schema class.
This is to avoid bringing in a dependency on Avro, since not
all plugins use Avro.

The second is that the RecordWritableConverter class was copied to
the elasticsearch module since that is the only place it is used,
and to avoid brining in a Hadoop dependency, since not all plugins
use Hadoop.

The third is that the deprecated classes were not copied over