CDAP-4950 add a workaround for the hive jdbc driver
Added a workaround for drivers that dont support commit, like the
hive jdbc driver. Added a plugin property for enabling auto
commit, which the input format sets to false by default. If auto
commit is enabled, any Connection created by drivers will be
wrapped with a NoOpCommitConnection that doesn't do anything when
commit is called.