Adding a DatabaseQuery action that can run after a pipeline run.
This allows people to do things like run a cleanup query at the
end of a pipeline run. Also adding the concept of macros to the
db configs and any other configs that want to use them. Macros
define a way to insert the formatted runtime into a config setting.
This allows somebody to set the db query to something like
'select * from table where date=${runtime:yyyy-MM-dd}', where
${runtime:yyyy-MM-dd} will be replaced by the runtime formmated as
year-month-day.