CDAP-5039 add macro support and change db plugins to use macros
Adding a MacroConfig class that has support for replacing certain
macros in string fields with some value based on the macro
function. Added runtime based functions that allow users to
calculate some date format using the logical start time of the run.

Currently only the database plugins make use of macros. There are
some limitations in the app that prevent transforms from using them,
namely the fact that the logical start time is not available to
transforms.