HYDRATOR-922 remove usage of BadRequestException
The class causes problems with some versions of spark that bundle
jersey classes in the spark-assembly jar. In those situations,
the javax.ws.rs classes visible to the program are the jersey
classes (which dont include BadRequestException) and not the
rs-api-2.0 classes (which do include BadRequestException). Since
that package seems to exist in multiple dependencies, just removing
usage of BadRequestException for simplicity.