CDAP-8132 add file path as an output field in file source
Adds an optional property for the file source to include the file
path that the record came from. The property cannot be used if
the input format class is set.

Also contains some refactoring with children of the FileBatchSource
to override config methods instead of some odd constructor json
string manipulation that was happening.