CDAP-14023 move file sink to core plugins
Moved hdfs sink to core-plugins as file sink, since it is able to
write to more filesystems than just hdfs. Also added the capability
to write files as avro or parquet in addition to just text.

Also created File versions of the HDFS move and delete actions
for the same reason. They are able to operate on file systems
other than just HDFS.