clean up hydrator-common
Removed unused Spark classes and removed the Spark dependencies.

Also removed all guava usage and the guava dependency. The common
module is used by many other projects, some of which may want to
use an incompatible guava version. It best to include as few
dependencies as possible for these types of modules.