CDAP-12852 add a property to copy headers of text files
This is a hack to allow wrangler's parse-as-csv to operate
correctly when a file is split into parts. Each record reader will
emit the header as its first record, which will allow an attached
wrangler transform to parse-as-csv with a header correctly.

The property is hidden, and is only expected to be set by the
dataprep service.