CDAP-14321 add formats to file source
Added csv, tsv, delimited, json, and blob formats to the file
source. This is to mirror the formats that are available in
the sink, so that whatever content is written by a File sink
can also be read by a File source.

Refactored much of the input format logic to be similar to what
is done for output formats. This mostly amounted to moving the
inner static record readers from PathTrackingInputFormat into
their own classes, and creating hooks to get the record reader
directly from the FileFormat enum.