CDAP-14470 make offset optional when reading with text format
Makes the offset field optional for the file source. It will be
there by default, but users can remove it if they don't want it.

This is also required to fix issues with wrangler, where it
makes assumptions about column numbering and ends up creating
bad data if the offset field is there.