CDAP-14481 check for extra schema fields
Usability improvements to the file source to check if there are
extra fields in the schema that are not supported when using the
text or blob formats. This is to address user behavior where
the format is changed but schema is not changed, resulting in a
invalid schema.

Also adding a get schema button to help users generate the correct
schema when changing formats.