use avro and parquet code for determining codec validity
Also fix a bug with tpfs and snapshot sinks where datasets created
by a pervious version of the plugin would not be readable, due
to differences in output format stored in the dataset properties.
Changed the sinks to override any properties from dataset creation
with those that the plugin knows are needed at runtime.