remove usage of StreamEvent
StreamEvent was a flowlet class and was being abused in the
FileStreamingSource. Removed usage of the stream based formats
and switched to more standard formats.

This is a backwards incompatible change, as clf, grok, and syslog
are no longer supported, but those formats weren't seeing much
use anyway.

Longer term, sources like this should find a way to hook into
the format plugins that were done for the batch file sources.