CDAP-15767 add support for reading int96 fields
Added support for reading int96 fields by copying relevant classes
from the parquet library and slightly modifying them to handle
int96 fields. INT96 fields can be read as bytes or as a long.
When read as a long, the value is converted into a unix timestamp
in microseconds, which means the nanosecond component of the
original data is lost.

Fixed a bug with the existing parquet format so that it correctly
uses the user provided schema as the read schema, which does not
have to be the same as the write schema.