CDAP-17607 add advanced join conditions
Added a condition type property that toggles between basic and
advanced modes. Basic is the currently behavior that joins on
key equality, while advanced is an arbitrary SQL expression.

Most of the implementation is simply to add configuration
properties and pass them down to the underlying framework.
Also did some cleanup on the JoinerConfig to make all the
properties consistent in style.