CDAP-19078 fix classloader for schema detection
now that filesystem traversal is happening in the connector and
in the file source instead of in the format plugins,
we should not be using a combine classloader
of the plugin classloader and the default classloader in the
Hadoop configuration. Instead, just the plugin classloader
should be used.

Without this, classes from the main classpath, like the s3
and gcs connectors, will be defined by the wrong classloader,
causing class cast exceptions.